Changed the sonogram/spectrogram generation code to use Mike's new generation and reduction algorithm.
Started implementing "Entities" as a means for users to organise their data.

Moved the jquery script reference into the master page. This is the only place that this reference should appear, due to probable changes to the Jquery prototype ($.fn) implementation from 1.2 to 1.3. Reloading the jquery script means that any custom wrappers added to the jquery prototype will be overwritten.