Changed Application.Exit event handler so (not working) web service is no longer called. A user's profile is now saved to isolated storage. Plan is to retrieve it when player loads again and send back to server (writing implemented, reading not implemented)
Gave the loaders a background color so tooltips work.
In Loader.cs fixed a few bugs. Introduced concept of maximumCount as a way of limiting the number of times a loader can be started.
Fixed various bugs with starting and stopping loaders (CompmediaElement, playlist.xaml.cs)

Implemented copy functionality for snippets. 
Incorporated JSON pretty printer for snippet copy and profile saver.

Adjusted when events and loaders are bound to CompMediaElement so proper loading and event response can occur.

Fixed (or rather supressed) a bug with the MeidaElement not playing to the very end of a chunk - made it best effort possible since the AudioReadingHandler often returns audio that is slightly shorter than requested.