Calculation of Summary and Spectral indices.
Testing and Debugging the calculation of Summary and Spectral indices.
Making sure that the appearance of the false colour spectrograms and the
tracks image of summary indices look OK and their information is
consistent with what is in the csv files.
testing on a variety of recordings with known and varied content.