Creation of new summary and spectral indices
Creation of new summary and spectral indices is now all located in the
same method. Now much easier - a) created a key; b) initialise
properties of new index in class IndexProperties; 3) calculate the index
in class IndexCalculate and store in class IndexStore.
To achieve the above I had to make changes which affected many other
classes.