Changed name of class IndexStore to IndexValues. Shifted all the static methods out of the class IndexProperties into a newly created class called InitialiseIndexProperties.
Also shifted all the index keys into the new class
InialiseIndexProperties.

The idea is to separate the properties of an index from its value. The
properties determine how a value will be normalised and displayed in its
relevant image.

Changes in all the other classes required to relocate the shifted string
keys