More work on SVD
created a new class within TowseyLibrary to do Singular Value
decomposition and also PCA.
At this stage the SVD methods have been tested on simple examples taken
from text books and are giving the correct answers. except that the
signs of the coefficients are frequently switched. Not sure what is
happening here. If the are switched in the U matrix they are also
switched in the VT matrix, so the reonstruction is not affected.