Worked on parallelisation bug
If the process runner suspends an executable that is taking too long, it
would previously not report any standard error or output. This behaviour
has now been modified so the std err and out are reported in the
exception (report actual program status at termination as opposed to
crashing silently on timeout with no info).