Completed tiling specced for website
Added new unit tests
Added functionality to the tiler to start drawing a super tile aligned
to the nearest tile boundary. WARNING: this use case is only for super
tiles that are not aligned to tile boundaries and are not proceeded by
another supertile in the layer. If these invariants fail tiles will be
overwritten (behaviour is undefined).