fixed calculation of EVN and indices
Fixed bug in calculation of the EVN spectral indices that arose due to
restricting the indexcalculation duration to 0.2s.
Also shifted calculation of indices to the newly created
IndexDistribution class.
Checked only the drawing of RGB LD spectrograms and the distribution
images.
Did not check the output from the Difference spectrograms which also
depend on the statistics.
Commented out a few lines in Xie's code to do with statistics on the
assumption that he does not use them!