Automated calc of Index Norm bounds
Experiment with different ways to automate the calculation of Index
Normalisation bounds which are used to prepare false colour
spectrograms. These norm bounds have a considerable effect on the
appearance of the final spectrogram image.
In the end decide to allow user to determine whether the bound for each
index should be be set by the user or should be calculated
automatically.