Fixed bugs in tiler and zooming spectrograms
Fixed bugs in the tiler for working with supertiles that were not
multiples of the super tile.
Added additional unit tests for the tiler.
Also changed the way ZoomTiledSpectrograms renders standard
spectrograms. A sliding window algorithm has been implemented so that
the Tiler always has adjacent supertiles to draw from when drawing tiles
that overlap multiple super tiles.