Finish work on Erroneous Indices
Finished work on code for warning user about segments of recording that
have produced erroneous index values.
At present there is only one error check for when three key indices sum
to ZERO. this situation arises when there is a FLAT or ZERO signal. This
condition is checked for when the indices are calculated but an vector
of zero index values is still returned.