Added a compromise to MasterAudioUtility
Allowing copying over previous files in super bad in production - it
would represent something wrong with the entire state of the system.

However in development, crashed programs running with the same args is
commonplace and not a sign of a world ending event.

Thus in RELEASE we ensure a harsher restriction.