Resurrect Spectral Diversity Index
Resurrect the Spectral Diversity Index which is based on an ART
clustering of the component spectra in a one-minute segment of audio.
Resurrected because it seems to perform well in a Susan Fuller
experiment and in one of mine.
To do this had to change a few other things as well!