Finding indeterminancy
Use three canetoad recordings to find indeterminancy in the recognizer
algorithm. Eventually trace it to the resampling process. The resampler
returns a slightly different array of WavReader.Samples for each read of
the .wav file. However if the file is at 22050 kHz then resampling not
required and the recognizer out put is deterministic i.e. exactly the
same values every run.