Concatenate Nature Conservancy files now working
In particular, it fills in missing portions of data.
For best results, the IndexPropertiesConfig file was edited slightly.
The Indonesian recordings have higher BGN and relatively less POW. Need
to think about best way to deal with this. Next job is to check that the
standard 24 hour concatenation is not broken.