Refactored various path related APIs
Turns out `AudioRecording`'s `FileName` property was always assigned the
result of `Path.GetFileNameWithoutExtension` and almost all usages of it
are wrong since they assume FileName has an extension (as is the normal
nonmenclenture of .NET) and do a duplicate file extension removal.

Thus I've renamed `FileName` to the more correct `BaseName`. I have not
fixed all instances of duplicate extension removal -- hopefully that is
something we can fix in time.

I also modified the `FileNameHelpers` API to clear up names of the
overloads. `AnalysisResultName` overloads only return a name, and
`AnalysisResultPath` overloads return a full path. Splitting this up
allows for a clearer API and also new useful overloads.