BREAKING CHANGE: VS2017 compatibility changes
This commit updates our projects to use new Visual Studio 2017 features
and fixes some bugs that were introdcued with Visual Studio 2017.

Due to  changes detailed here:
https://developercommunity.visualstudio.com/content/problem/15434/cannot-find-wrapper-assembly-for-type-library-envd.html
The EnvDTE debugging hack for automatically attaching to console
instances of AnalysisPrograms.exe was broken. Thus this is a breaking
change for your build - unless VS2017 is installed you won't have the
DLLs in the correct places!