Converted LdSpectrogramConfig.XAxisTicInterval to an integer
There were several spots in the code were dersializing XAxisTicInterval
was failing or behaving weird because TimeSpan assumes that intergers
are days when parsing strings!

Because our convention is that integers for time units should be
seconds, I've changed XAxisTicInterval to an int, and updated the code.

Running the unit tests revealed the need to sometime support parsing old
JSON data files where XAxisTicInterval is properly encoded as a
timepsan. These cases are now supported by LagacyTimeSpanDataConverter
class.