Testing CSV serializer wrt to TimeSpans
I changed the CSV Serializer so that b default it outputs better error
messages for bad type conversions. I also added a unit test to make sure
that it works. An important and unfortunate side effect is that we had
to disable yield the CSV stream. It remains to be seen if this has a big
impact on performance.

Additionally I wrote a unit test to ensure that the CSV Serializer is
round tripping TImeSpan values correctly. The test tests 30 orders of
magnitude for randomly generated time spans. This is necessary because
an earlier version of .NET changed the default formatting for TimeSpans
which sometimes causes serialization errors. This test ensures our
serializer is not at fault and that new data generated is not at fault
either.