More work on LdSpectrogramConfig.XAxisTicInterval
The changes in 4faf291c6955a2b4f206c53ddfc237417612475b had a bug in
them that incorrectly parsed the tick interval as seconds instead of
hours.

I've reworked the solution so that serialization is done as a double in
seconds and the timespan conversion happens in the LdSpectrogram class.
This solution is simpler: instead of actively trying to maintain
compatibility with old datasources (where XAxisTicInterval is stored as
JSON) those cases will now simply just use the default of an hour (which
almost always was the default anyway).

I've also updated the config files to use the new property.