EventStatistics analysis almost complete!
It now runs sucessfully most of the time.
Just got to iron out a few runtime bugs.

Also refactored AudioRecording.Duration() to a getter property. This is
only a cosmetic change.

Renamed EventBase.MinHz to LowFrequencyHertz to be consistent with other
areas of our work. THIS IS A BREAKING CHANGE