Refactored CSV module
Simplified the behaviour of TwoDimensionalArray enum. As a description of element ordering it was very confusing. Changing it to the concept of a transform that is applied is much easier to comprehend. Luckily it only needed to be changed in a few spots because there's only one function (ReadSpectralIndices) that uses DecodeMatrix.

Also refactored to support better naming and Zio file inputs for some cases.