More work on Zooming Spectrogram
Analysis IO now has a subclass that ensure the input is a directory. Future work will add file equivalent.

Changed the naming format of the tiles to be closer to ISO8601.

Allowed filename helper to accept an empty basename for when we intentionally want to not prepend basename to a result file. Useful for results that are stored in a package (that is named with the basename).