Migrated SqliteFileSystem to open source package
SqliteFileSystem started out as an experiment but was useful enough to split out into it's own package as an open source project.

This commit removes the local copy of the code and updates the solution to use the nuget package.

Please note: the stand alone package is not yet published on nuget so we are using the Appveyor nuget feed to host the packages.

This is work done for #120