Fixes poor peformance on small machines
The TPL Parallel.For is designed for fien grained parallelism. To that end it librerally uses the all available threads for processing units of work. This is fine usually but the workloads inside of AnalysisCoordinator are very big. When too many threads all do the same large work item (in my case, 64 threads generating high res indices on a 16GB RAM machine) huge amounts of RAM are required, the process pages, and effectively stalls.

This patch limits the concurrency in AnalysisCoordinator to the number of CPU cores available and actually increaseces processing speed on most small machines. In my example task the proceses finishes in a few hours, as opposed to never finishing after 30+ hours.

Unfortunately I can not figure out a way to tune the parallelism for high specced machines.