Fixed bug with naming of summary indices CSV
When the concatenate indices code ran, it output the Summary Indices with a differant name format than the standard indices output. This would make the output of the concatenate code invalid indices!

I've fixed it so that the same name is used for the file both when generating indices and when concatenating.

I've done some refactoring as well to reduce duplicate code and to centralize strings (getting rid of magic strings).