[WIP] Removing dynamic yaml
This commit is the bulk of the work required to remove DynamicYaml configuration and replace it with an extendable generic config file, which can also be used for static typing.

Note, build fails, tests fail. More work needed.

Note, nuget also needed to a package reinstall. It looks like the stylecop analyzers might be broken.

By Adding a `Parse` method to the `IAnalyser2` interface we can allow for analyzers to parse their own config files to static classes :grinning:. However, this means the way we resolve Analyses by identifiers has changed and is now dependent on either an explicit argument, or the first two sections (separated by a period) of a config file name.

This work was done for https://github.com/QutEcoacoustics/audio-analysis/issues/119