Completes removal of DynamicYaml
This is a continutation of 72aab48d1488622e535d147de521512691102a5a which due to issues with the build could not be tested.

This commit runs ensures all the unit tests works and additionally simplifies the `Config` object signficantly (it no longer relies on scanning YAML document nodes but rather uses a generic deserialization to store an object graph). We've also moved to using a `Loaded` event to fire post load hooks; the previous method that coopted property setters was ultimately not useful because it was invoked during serialization, before all needed properties were set.

As we use American spelling for code:
 - this commit to fixes the spelling on `Yaml.Deserialize`.
 - this commit also fixes the spelling of Neighborhood.

Closes #119