Fixes (most of) the unit tests
The work done in 864f7a491e2ea0e938161bd390c1c931ecbdf63c to reorganize the solution for our open source release (see #140) was left incomplete because various dependencies needed to be replaced.

This commit fixes nearly all tests (with the exception of one test which is now detailed in #146).

This commit also:
- Replaces our chosen cloning library (GerorgeCloney) with a newer one that can clone `FileInfo`s via reflection
- Fixes a bug from 72aab48d1488622e535d147de521512691102a5a where `FrequencyScaleType` property did not match the name used in the config files
- The changed FFT calculation produced very very small changes in anything that depends on the FFT. I either recalculated our expected binary matrices, or for most failing cases added a delta threshold to the comparison. I found 1E-9 was plenty of delta and three orders of magnitude smaller than any change we could notice
- I adjusted the build to ensure audio tools are copied to test output