Fixed bugs with FakeAnalysis
Fake analysis used async functions that behave differently on Mono!!! It looks like Mono does not wait for async void tasks to complete, unlike in the .NET platform. Modified the analysis so that it always returns a value.

Also fixed a our custum InRange attribute which was built on the incorrect premise that it would recieve a double (after parsing) but it in fact recieves a string.