Fix bug with wavunpack bug on Linux
wvunpack was failing to extract .wav segments on Linux; it was attempting to read the second path as an input file, rather than assuming it was an output file.

Version 4.31 of wavpack introduced the special "-o" argument on linux only builds to allow multiple input files. This change was not mirrored on Windows (and crashed wavpack until version 4.50).  This changes adds the "-o" option for both platforms so that Linux runs now work correctly.

Our servers have got fresh versions of wavunpack (version 5.0)  so the bundled Windows version was updated as well.