Updates CLI lib
We were using a custom build of the new CLI lib. The lib has been updated with the missing features we need so this commit updates to the official build.

This is a checkpoint commit. We found a bug while making this change and we're waiting for that bug to be merged in.