Converted files to git lfs
Many of the fixtures in the tests folder matched rules that stated the files in the the fixtures should be tracked as LFS files. Because they were ordinary files (and because git lfs changed it behaviour) these files were automatically converted on checkout. The result was a dirty repo on our build server which falsely set the DIRTY flag in the metadata of our build.