Another round of improvements for ProcessRunner
More stabs in the dark.

The changes in 72cc826d6f3e1008b95fc6e511af8041099d1273 did not have any affect on the timeouts, however the processes that time out DO now capture their standard out and error. This means that even though processes are timing out, they are tantilizingly close to actually working.

This change gives the process runner a chance to return as if it were sucessful, despite timing out.

We also improve the logging so we can track the effects on the concurrency better.