Adds a Dockerfile
Adds a Dockerfile. This image produced by this docker file is based on debian, has all of our required dependencies installed (including audio tools and Mono). It also has PowerShell installed. The docker file also contians the binary assets for a given version.

This commit also includes a script to download binary assets from either AppVeyor or Github. That script is included in the docker container so users can update the version of AP.exe they use when running the container.

Provisional support is added for building these images when appveyor does CI builds, but I expect changes will need to be made.