Fixes a bug in EventStatistics
This patch ensures that the audio recording duration used by event statistics is always rounded down to the nearest second. There is a subtle conflict in logic later on in the event downloading  code that requires audio segments be rounded to the nearest second. If the recording duration is greater than 0.5, and the event ends at the end of the recording, then the requested media is rounded up to the next second, past the end of the recording. This patch prevents that error.