Fixes bad parsing of values from audio tools
For computes with a default culture where the default radix point character is not a `.` (a period) the parsing of the duration and other fields from SoX and FFmpeg were erroneous.

The fix is to ensure we use invariant culture for parsing the duration reported from SoX.

Fixes #180.

We also add some additional logging to the segment calculator to help debug potential future problems.