Adds support for very long audio recordings
Fixes #178

Both FFmpeg and SoX audio utilities now report durations of audio for very long files (greater than 24 hours). We should additionally find the precision for durations of audio recordings will have been improved. However, it should be noted that storage in TimeSpans obfuscates most of this improvement.

The SoX audio utility was also changed so that the second invocation that used the `-D` switch now only happens if needed. This should result in a minor performance boost. I also allowed SoX to process files directly. This may need to be reverted if it proves problematic.