FileSegment now produces less warnings
Fixes #174 .

Split the FileSegment constructor that supports taking explicit information into two parts: one for taking a known date (which is used for Split and Clone) and one that still requires a date to try and be parsed from the filename. This means that when the segment is split, it will no longer try and parse the file date again and wont spam N+1 warning messages.

Also did a little bit of refactoring and improved the warning message output when an ambiguous date is parsed.