Adds auto attach feature for DEBUG builds
Now if the env var AP_AUTO_ATTACH is set (in DEBUG builds) AP.exe will automatically try and attach the VS debugger.

This can be useful if you want the debugger to attach before CLI argument parsing is done. This feature formalizes a hacky comment that used to be in MainEntry.cs