[BREAKING] Refactored logging setup
We have removed logging configuration via the log4net config and instead opted for an entirely procedural logging setup. This has a few advantages: it is easier to understand (IMO); it centralizes the logging setup; it makes it easier to customize the logging setup for unit testings; it means that we setup the logging in a smarter way (we don't need to relay on filters to sort log messages, we can instead send log messages directly to our desired loggers/appenders); and finally it simplifies post-initialization configuration (as done by CLI options) since we have object references to all of the loggers/appenders.

BREAKING change: log files are now suffixed with a datetime stamp of the format: "log_20180723T001942Z.txt". This change is important for one major reason: this unblocks running multiple instances of AP.exe from the same copy of the executables. Previously if two instances of AP.exe ran at the same time, the logger would conflict with both instances trying to write/lock the log.txt file. This no longer happens, and if perchance an instance starts at the exact same second, numeral suffixing prevents file name conflicts.

We should also see very slight improved performance (e.g. at most 100 disk I/O operations are skipped) for log file cleaning as every previous log no longer needs to be renamed (a behaviour of the RollingLogFileAppender) whenever AP.exe is started.