Now all config files are dumped to the log
This is an important change that helps ensure that experiments are 100% reproducible. With this change, all input to program (source, arguments, configuration, and program versioning) are now logged so that an execution should be reproducible and unambiguous.

This change also has the useful side effect of ensuring all config file reading is now cached (which should lower I/O calls). All cached config objects are cloned before being returned to ensure cross-thread immutability.

IndexProperties are now loaded through the ConfigFile class for consistency (and provenance).

Closes #133.