Fixes a bug in chromeless LDFC images
It seems at some point the code used to draw frequency lines went from creating its own copy of the image to draw on, to modifying the source image. This in turn modified the "clean" image we were using for the chromeless function.

I've fixed the bug by cloning the image sent to the spectrogram framing function and written a unit test for it.