Changes to allow calculation of OSC index
Needed to make changes to Index calculate to add this to list of spectral indices

Attending Issue #146 Oscillations test

This commit and the previous 7 commits have all to do with Issue #146.
In this commit set up the IndexCaltulate test to include the new OSC spectral index.
This issue #146 should now be closed except that we still hvae to check how the index is working to produce LDFC spectrograms.