Fixes bug in CompressIndexSpectrograms
The changes in 75eaf71efe52ff3c245eb9ae56e45c2f7aede75b uncovered an existing bug in CompressIndexSpectrograms where a fixed size array was always used to store a window of values, even when the the window slid past the end of the array. The result was any tail values were left as 0.0 and threw off the statistics.

Closes #152