Unit test fixes for rendering all gray FCS
Two of the failing tests were just due to different numbers of files being produced.

The other failed in an aritficial data scenario where the min and max of the matrix were equal. The resulting range (0) caused NaNs  when normalized. This bug was discovered now due to the dual auto-normalizations changes to the index properties config for R3D.

Fix is to revert to pre-defined bounds is auto-noramlization range is 0.

This fixes work done in 6ec16c68ae7ceccb1595c2cf624e9aa94195cf3e for #199