Fixes CLI duplicate value parser bug
The fix introduced in 29175199ade618bc601f4d691065e9ad5d0b88cc resulted in AP.exe crashing every time it was run.

This patch ensures our value parsers are used, and more so adds some sorely needed tests for our MainEntry. These tests are useful in their own right, but for this bug, they also ensure CommandLineUtils runs and does it's validation. Previously, because CLUs validation was never executed in our tests it was possible (and happened) to ship a version of AP.exe that passed all tests but wouldn't even start!