Adds checkes for critical DLLs
Now does a rather paranoid check everytime AP.exe starts to ensure we can load System.ComponentModel.DataAnnotations. Because this DLL was in the GAC and because our dependencies depend on it, it was a rather insidious bug to fix. The only way to check for it was to do so before other things depend on it.

Closes #225