Refactor code which draws histograms
Use the full range of the y-axis scale when drawing histogram.
Add the number of the modal bin to the collected statistics.
Change method arguments to accept modalBin.
Remove duplicated lines from SpectrogramConstants.cs.
Write SandPit method to test drawing of histograms.
this can be turned into a unit test.
NOTE: The "gap" in the histograms noted in Issue #186 is due to way Rectangular bars were drawn. Nothing to do with inadequate size of int.