Changed way mode of distribution calculated
One reason for suboptimal rendering of LFFC spectrograms is that in some cases the mode of the distribution of index values is also the minimum, expecially in case of Entropy where mode often = zero. When tihs occur, now prevent mode being in the bottom two histogram bins.
Most of the other changes are resharper inspired or designed to simplify code.