Experiments with concatenate
Previous commits have helped fix this problem but not completely. Experimental runs on the HPC revealed a bug in calculation of 98th percentile for CVR and EVN. A hack/fix for this is inserted into IndexDistributions.cs at line 77.
Insert option in LDSpectrogramstitching.cs for different search option. I need it for my setup.