Fixes culture formatting bug for SoX commands
In some cultures, where a comma is used as the decimal separator. In those cases, when we format commands for SoX the numbers would be separated with commas and the command would be invalid.

The commit fixes that by ensuring we only use the invariant culture when formatting floating point types.

Fices #242