Bumped System.Numerics.Vectors and ImageSharp
Fixes #241

Due to some complicated diamond dependency problems I do not really understand, there was a conflict with the version of System.Numerics.Vectors. ImageSharp required a version shipped only in .NET Standard, which we did not have access to.

The only solution I could find was to move to the larest dev build of ImageSharp, which changed this dependency.

This change includes app binding redirects, ImageSharp API updates, and a few tweaks to the reflection type load exception formatter so that it now outputs fusion logs.

Still need to check that tests pass.