Adds test for BARLT metadata parsing
FLAC custom metadata fields are output as is by ffprobe. Because they did not have an `=` character the `Substring` call in `GetInfo` threw an exception because the `IndexOf` call returned `-1` and that was not a valid index for a string.

I've added test audio for a BARLT and a test dedicated to parsing that metadata. At this stage we have no actual use for the metadata.