Add path diagnostics method
The path diagnostics method will do it's best to diagnose why a given path might not exist. It will detect paths that are too short, or that have spaces in folder names, where errors occurs, and will even make suggestions.

I've added this to help diagnose failing tests on the CI, but it will also make friendlier error messages if used with CLI arguments.