Fix more tests
Add configuration to build metadata so unit tests actually look in correct build folder for files.

Replace debug.writeline with trace.writeline so that
console output is available for release mode tests
as well.

Make master audio utility constructor more resiliant now that we no
longer wvunpack to be present in a system.

Work done for #196