Rework the old SpectralTrack class.
Issue #297 Remove contents of old class and repurpose for keeping path of a spectral track.
The old class was not being used.
The new class is now used by VerticalTrackParameters and will be used by SpectralPeakTrackParameters.

Changes in Image_MultiTrack.cs are resharper inspired. But this class was the only one to use the old SpectralTracks.