Fixes faulty LFS file check error
We need files to be restored from LFS before we can use them - otherwise they are empty!

The previous check I had could not distinguish between an empty file and file that had been modified.

I've now written a far more detailed script that should be able to handle new and modified files.

It also now skips running if it was run withing the last hour