Makes working with test result artefacts easier
If you use SaveTestOutput (a method on OutputDirectoryTest) a saved file will be saved to it's individual directory (as before), registered with mstest, and now (NEW) will also be saved in a daily test output folder in tests/Results/.

This flattened-folder output makes it much easier to compare output between tests when we're debugging.

We have to adhere to the mstest results directory layout for CI tests so SaveTestOutput makes a copy of the file to daily output.

Daily output is disabled for the CI server