Update CommonParameters.cs
Issue #370 Change double to array of double. THis is with a view to running the call recognizers over a set of decibel thresholds rather than just one threshold.
This is to accomodate calls that are close and distant to the microphone, i.e. to accomodate a range of call SNR.