Fixed bug that arose when first implementing multiple decibel thresholds
Issue #390 A bug appeared when first implementing multiple decibel thresholds. The loop over an array of thresholdholds was implemented only around the profile matching and did not include the post processing. This caused the recognizer to miss many true calls. The corrections implemented here are to encusre that the lopp over all decibel thresholds includes both the profile detection and the post processing in each pass.