Changes to the sideband activity filter
Issue #390 I changed the way the event filter is working that filters on acoustic activity in the sidebands. I changed the name of a variable and its function. The sidebnad filter now checks for two kinds of sideband activity - the average background decibels in the sidebands, and it looks for brief events in the sidebands.