Removed `Tag` from AcousticEvent
The Tag property was a poorly implemented method of tagging an acoustic event with a boolean value.

Unfortunately we deal with textual tags far more freuqently and the usage is confusing - particularly if the field is detected by automated tooling in output results.

It was only used by two methods that calculated accuracy. Neither of those methods are curerntly used.

Added to all that, AcousticEvent is on the deprecation path... so this change should be relatively minor